# services/api-gateway/docs/swagger.yaml
basePath: /
definitions:
  gin.H:
    additionalProperties: {}
    type: object
  handlers_admin.employeeLoginRequest:
    properties:
      email:
        type: string
    type: object
...
  handlers_auth.requestCode:
    properties:
      phone:
        type: string
    type: object
...
  handlers_auth.verifyCode:
    properties:
      code:
        type: string
      phone:
        type: string
    type: object
...
  model.MaterialsAndServices:
    properties:
      id:
        type: integer
      item:
        type: string
      quantity:
        type: integer
      visitID:
        type: integer
    type: object
  model.Patient:
    properties:
      birthDate:
        type: string
      email:
        type: string
      firstName:
        type: string
      gender:
        type: string
      password:
        type: string
      phone:
        type: string
      secondName:
        type: string
      surname:
        type: string
      user_id:
        type: integer
    type: object
...
  model.UnconfirmedAppointment:
    properties:
      created_at:
        type: string
      date:
        type: string
      doctor:
        type: string
      gender:
        type: string
      id:
        type: integer
      patient_birth_date:
        type: string
      patient_first_name:
        type: string
      patient_id:
        type: integer
      patient_second_name:
        type: string
      patient_surname:
        type: string
      phone_number:
        type: string
      status:
        type: string
      time:
        type: string
      updated_at:
        type: string
    type: object
...
  model.UpdateAppointment:
    properties:
      date:
        type: string
      id:
        type: integer
      status:
        type: string
      time:
        type: string
      updated_at:
        type: string
    type: object
...
  model.VisitPayment:
    properties:
      created_at:
        type: string
      doctor:
        type: string
      materials_and_services:
        items:
          $ref: '#/definitions/model.MaterialsAndServices'
        type: array
      patient:
        type: string
      price:
        type: integer
      visit_id:
        type: integer
    type: object
  model.VisitPaymentUpdate:
    properties:
      price:
        type: integer
      status:
        type: string
      visit_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: ПО для стоматологической клиники
  title: Примастом API-Gateway
  version: "1.0"
paths:
...
  /api/admins/{id}:
    delete:
      description: Удаляет пользователя по ID
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Неверный ввод
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      summary: Удалить пользователя
      tags:
      - Администратор
...
  /api/completed-visits:
    get:
      description: Возвращает список неоплаченных посещений с материалами и услугами
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.VisitPayment'
            type: array
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      summary: Получить список неоплаченных посещений
      tags:
      - Администратор
  /api/completed-visits/{id}:
    put:
      consumes:
      - application/json
      description: Обновляет информацию об оплате посещения по ID
      parameters:
      - description: ID посещения
        in: path
        name: id
        required: true
        type: integer
      - description: Данные об оплате
        in: body
        name: payment
        required: true
        schema:
          $ref: '#/definitions/model.VisitPaymentUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Неверный ввод
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      summary: Обновить оплату за посещение
      tags:
      - Администратор
...
  /api/doctors/{id}:
    delete:
      description: Удаляет пользователя по ID
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Неверный ввод
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      summary: Удалить пользователя
      tags:
      - Администратор
...
  /api/patients-login/{id}:
    put:
      consumes:
      - application/json
      description: Обновляет логин сотрудника по ID
      parameters:
      - description: ID сотрудника
        in: path
        name: id
        required: true
        type: integer
      - description: Новый логин
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/handlers_admin.employeeLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Неверный ввод
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      summary: Обновить логин сотрудника
      tags:
      - Администратор
  /api/patients/{id}:
    delete:
      description: Удаляет пользователя по ID
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Неверный ввод
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      summary: Удалить пользователя
      tags:
      - Администратор
    put:
      consumes:
      - application/json
      description: Обновляет информацию о пациенте по ID
      parameters:
      - description: ID пациента
        in: path
        name: id
        required: true
        type: integer
      - description: Новые данные пациента
        in: body
        name: patient
        required: true
        schema:
          $ref: '#/definitions/model.Patient'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Неверные входные данные
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      summary: Обновить данные пациента
      tags:
      - Администратор
...
  /api/unconfirmed-appointments:
    get:
      description: Возвращает список неподтверждённых записей на приём
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.UnconfirmedAppointment'
            type: array
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      summary: Получить неподтверждённые записи
      tags:
      - Администратор
  /api/unconfirmed-appointments/{id}:
    put:
      consumes:
      - application/json
      description: Обновляет дату, время и статус записи по ID
      parameters:
      - description: ID записи
        in: path
        name: id
        required: true
        type: integer
      - description: Новая информация о записи
        in: body
        name: appointment
        required: true
        schema:
          $ref: '#/definitions/model.UpdateAppointment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Неверный ввод
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      summary: Обновить запись на приём
      tags:
      - Администратор
...
  /auth/request-code:
    post:
      consumes:
      - application/json
      parameters:
      - description: Телефон
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers_auth.requestCode'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Запрос кода подтверждения по телефону
      tags:
      - auth
      - Авторизация
  /auth/verify-code:
    post:
      consumes:
      - application/json
      parameters:
      - description: Телефон и код
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers_auth.verifyCode'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Подтверждение кода, отправленного по телефону
      tags:
      - auth
      - Авторизация
...
swagger: "2.0"